Make the length cast explicit
This is could be a bug if the len is greater than INT64_MAX, but the
code doesn't seem to care about this, so adding the explicit cast
to remove the compiler warning.

+++ src/civetweb.c
@@ -3183,7 +3183,7 @@ int mg_write(struct mg_connection *conn, const void *buf, size_t len)
 		}
 		allowed = conn->throttle - conn->last_throttle_bytes;
 		if (allowed > (int64_t)len) {
-			allowed = len;
+			allowed = (int64_t)len;
 		}
 		if ((total = push(NULL,
 		                  conn->client.sock,
